+++ app/models/volunteer.py
@@ -1,7 +1,10 @@
 # standard library
 import datetime
 import itertools
-from typing import Any
+import random
+import time
+import string
+from typing import Any, Generator
 
 # third party library
 import pandas as pd
@@ -10,9 +13,11 @@ from sqlalchemy import Column, Integer, Text, DateTime, Float
 # local library
 from app import db
 from app.constants import MODEL_PRETRAINED
+from app.models import BaseDataModel
+from app.tools.generate import truncated_gauss
 
 
-class Volunteer(db.Model):
+class Volunteer(db.Model, BaseDataModel):
     __tablename__ = "volunteer"
 
     # other_columns
@@ -90,12 +95,12 @@ class Volunteer(db.Model):
         Parameters
         ----------
         + `form` (dict[str, Any]) : 表單
-            - `key` (str) : `Volunteer.data_columns` 與 `Volunteer.other_columns` 中所有的名稱
+            - `key` (str) : 只將存在於 `Volunteer.data_columns` 與 `Volunteer.other_columns` 的欄位設值
             - `value` (Any) : 對應值
 
         Caution
         -------
-        初始化時，沒有設的欄位預設就是 None。
+        初始化時，沒有設置的欄位預設為 None。
         初始化後，需再進行預測，方可填入資料庫。
 
         ```py
@@ -122,7 +127,7 @@ class Volunteer(db.Model):
         return pd.Series(self.jsonify()).to_string()
 
     def jsonify(self) -> dict[str, Any]:
-        """轉為 JSON 格式
+        """轉為類 JSON 物件
 
         Returns
         -------
@@ -157,9 +162,47 @@ class Volunteer(db.Model):
     def predict(self) -> None:
         """將實例丟入模型預測
 
-        將 `Volunteer.result_columns` 裡的名稱將設置為屬性，其值為預測結果
+        將 `Volunteer.result_columns` 中的欄位設值 (預測結果)
         """
         data = self.standardize()
         result_probas = map(float, MODEL_PRETRAINED.predict_proba(data)[0])
         for column, result_proba in zip(Volunteer.result_columns, result_probas):
             setattr(self, column, result_proba)
+
+    @staticmethod
+    def generate(n: int) -> Generator["Volunteer", Any, None]:
+        """生成 n 筆隨機資料
+
+        Parameters
+        ----------
+        + `n` (int)
+        """
+        for _ in range(n):
+            volunteer_dict = {
+                "buildDate": str(int(time.time() * 1000)), # unixtime (ms)
+                "HighBP": str(random.randint(0, 1)),
+                "HighChol": str(random.randint(0, 1)),
+                "CholCheck": str(random.randint(0, 1)),
+                "Smoker": str(random.randint(0, 1)),
+                "Stroke": str(random.randint(0, 1)),
+                "HeartDiseaseorAttack": str(random.randint(0, 1)),
+                "PhysActivity": str(random.randint(0, 1)),
+                "Fruits": str(random.randint(0, 1)),
+                "Veggies": str(random.randint(0, 1)),
+                "HvyAlcoholConsump": str(random.randint(0, 1)),
+                "AnyHealthcare": str(random.randint(0, 1)),
+                "NoDocbcCost": str(random.randint(0, 1)),
+                "DiffWalk": str(random.randint(0, 1)),
+                "Sex": str(random.randint(0, 1)),
+                "BMI": str(truncated_gauss(15, 35, sigma=3)),
+                "Age": str(truncated_gauss(1, 3, sigma=3)),
+                "Education": str(truncated_gauss(1, 6, sigma=1)),
+                "Income": str(truncated_gauss(1, 11, sigma=2)),
+                "GenHlth": str(random.randint(1, 5)),
+                "MentHlth": str(random.randint(0, 30)),
+                "PhysHlth": str(random.randint(0, 30)),
+                "EmailAddress": "".join(random.choice(string.ascii_letters) for _ in range(16)) + "@gmail.com"
+            }
+            volunteer = Volunteer(volunteer_dict)
+            volunteer.predict()
+            yield volunteer